+++ TestDrivenDev/TDD_PivotStructure/PivotStructure/Properties/AssemblyInfo.cs
