+++ TestDrivenDev/TDD_PivotStructure/PivotStructure/src/Tests/Attributes/FieldMapping.cs
